Remove various legacy code paths
