+++ src/Symfony/Bundle/SecurityBundle/DependencyInjection/MainConfiguration.php
@@ -205,11 +205,6 @@ class MainConfiguration implements ConfigurationInterface
             ->scalarNode('provider')->end()
             ->booleanNode('stateless')->defaultFalse()->end()
             ->scalarNode('context')->cannotBeEmpty()->end()
-            ->booleanNode('logout_on_user_change')
-                ->defaultTrue()
-                ->info('When true, it will trigger a logout for the user if something has changed. Note: No-Op option since 4.0. Will always be true.')
-                ->setDeprecated('The "%path%.%node%" configuration key has been deprecated in Symfony 4.1.')
-            ->end()
             ->arrayNode('logout')
                 ->treatTrueLike([])
                 ->canBeUnset()
@@ -415,10 +410,6 @@ class MainConfiguration implements ConfigurationInterface
                             ->end()
                             ->scalarNode('memory_cost')->defaultNull()->end()
                             ->scalarNode('time_cost')->defaultNull()->end()
-                            ->scalarNode('threads')
-                                ->defaultNull()
-                                ->setDeprecated('The "%path%.%node%" configuration key has no effect since Symfony 4.3 and will be removed in 5.0.')
-                            ->end()
                             ->scalarNode('id')->end()
                         ->end()
                     ->end()
