+++ src/Symfony/Component/Config/Definition/Builder/ParentNodeDefinitionInterface.php
@@ -15,8 +15,6 @@ namespace Symfony\Component\Config\Definition\Builder;
  * An interface that must be implemented by nodes which can have children.
  *
  * @author Victor Berchet <victor@suumit.com>
- *
- * @method NodeDefinition[] getChildNodeDefinitions() Gets the child node definitions - not implementing it is deprecated since Symfony 4.2
  */
 interface ParentNodeDefinitionInterface extends BuilderAwareInterface
 {
@@ -43,4 +41,11 @@ interface ParentNodeDefinitionInterface extends BuilderAwareInterface
      * @return $this
      */
     public function append(NodeDefinition $node);
+
+    /**
+     * Gets the child node definitions.
+     *
+     * @return NodeDefinition[]
+     */
+    public function getChildNodeDefinitions();
 }
