+++ src/Symfony/Component/HttpKernel/KernelInterface.php
@@ -21,8 +21,6 @@ use Symfony\Component\HttpKernel\Bundle\BundleInterface;
  * It manages an environment made of application kernel and bundles.
  *
  * @author Fabien Potencier <fabien@symfony.com>
- *
- * @method string getProjectDir() Gets the project dir (path of the project's composer file) - not defining it is deprecated since Symfony 4.2
  */
 interface KernelInterface extends HttpKernelInterface
 {
@@ -129,6 +127,13 @@ interface KernelInterface extends HttpKernelInterface
      * @deprecated since Symfony 4.2
      */
     public function getRootDir();
+
+    /**
+     * Gets the project dir (path of the project's composer file).
+     *
+     * @return string
+     */
+    public function getProjectDir();
 
     /**
      * Gets the current container.
