+++ src/Symfony/Component/Ldap/Adapter/EntryManagerInterface.php
@@ -16,22 +16,14 @@ use Symfony\Component\Ldap\Exception\LdapException;
 use Symfony\Component\Ldap\Exception\NotBoundException;
 
 /**
- * Entry manager interface.
- *
  * @author Charles Sarrazin <charles@sarraz.in>
  * @author Bob van de Vijver <bobvandevijver@hotmail.com>
  * @author Kevin Schuurmans <kevin.schuurmans@freshheads.com>
- *
- * The move() methods must be added to the interface in Symfony 5.0
- *
- * @method void move(Entry $entry, string $newParent) Moves an entry on the Ldap server
  */
 interface EntryManagerInterface
 {
     /**
      * Adds a new entry in the Ldap server.
-     *
-     * @param Entry $entry
      *
      * @throws NotBoundException
      * @throws LdapException
@@ -41,26 +33,29 @@ interface EntryManagerInterface
     /**
      * Updates an entry from the Ldap server.
      *
-     * @param Entry $entry
-     *
      * @throws NotBoundException
      * @throws LdapException
      */
     public function update(Entry $entry);
 
     /**
+     * Moves an entry on the Ldap server.
+     *
+     * @throws NotBoundException
+     * @throws LdapException
+     */
+    public function move(Entry $entry, string $newParent);
+
+    /**
      * Renames an entry on the Ldap server.
      *
-     * @param Entry  $entry
-     * @param string $newRdn
-     * @param bool   $removeOldRdn
+     * @throws NotBoundException
+     * @throws LdapException
      */
-    public function rename(Entry $entry, $newRdn, $removeOldRdn = true);
+    public function rename(Entry $entry, string $newRdn, bool $removeOldRdn = true);
 
     /**
      * Removes an entry from the Ldap server.
-     *
-     * @param Entry $entry
      *
      * @throws NotBoundException
      * @throws LdapException
