+++ src/Symfony/Component/Ldap/CHANGELOG.md
@@ -1,16 +1,22 @@
 CHANGELOG
 =========
 
+5.0.0
+-----
+
+ * Added method `move() to `EntryManagerInterface`
+ * Added pagination support to the ExtLdap adapter with the pageSize query option
+
 4.3.0
 -----
 
- * added `EntryManager::move`, not implementing it is deprecated
+ * Added `EntryManager::move`, not implementing it is deprecated
  * Added pagination support to the ExtLdap adapter with the pageSize query option
 
 4.2.0
 -----
 
- * added `EntryManager::applyOperations`
+ * Added `EntryManager::applyOperations`
  * Added timeout option to `ConnectionOptions`
 
 4.1.0
@@ -22,13 +28,13 @@ CHANGELOG
 4.0.0
 -----
 
- * removed the `LdapClient` class and the `LdapClientInterface`
- * removed the `RenameEntryInterface` interface and merged with `EntryManagerInterface`
+ * Removed the `LdapClient` class and the `LdapClientInterface`
+ * Removed the `RenameEntryInterface` interface and merged with `EntryManagerInterface`
 
 3.3.0
 -----
 
-* The `RenameEntryInterface` inferface is deprecated, and will be merged with `EntryManagerInterface` in 4.0.
+ * The `RenameEntryInterface` inferface is deprecated, and will be merged with `EntryManagerInterface` in 4.0.
 
 3.1.0
 -----
