+++ src/Symfony/Component/Security/Core/Encoder/PasswordEncoderInterface.php
@@ -17,8 +17,6 @@ use Symfony\Component\Security\Core\Exception\BadCredentialsException;
  * PasswordEncoderInterface is the interface for all encoders.
  *
  * @author Fabien Potencier <fabien@symfony.com>
- *
- * @method bool needsRehash(string $encoded)
  */
 interface PasswordEncoderInterface
 {
@@ -47,4 +45,9 @@ interface PasswordEncoderInterface
      * @throws \InvalidArgumentException If the salt is invalid
      */
     public function isPasswordValid($encoded, $raw, $salt);
+
+    /**
+     * Checks if an encoded password would benefit from rehashing.
+     */
+    public function needsRehash(string $encoded): bool;
 }
