+++ src/Symfony/Component/Security/Core/Exception/AuthenticationException.php
@@ -59,25 +59,6 @@ class AuthenticationException extends RuntimeException
     }
 
     /**
-     * {@inheritdoc}
-     *
-     * @final since Symfony 4.3, use __serialize() instead
-     *
-     * @internal since Symfony 4.3, use __serialize() instead
-     */
-    public function serialize()
-    {
-        $serialized = $this->__serialize();
-
-        if (null === $isCalledFromOverridingMethod = \func_num_args() ? \func_get_arg(0) : null) {
-            $trace = debug_backtrace(DEBUG_BACKTRACE_PROVIDE_OBJECT, 2);
-            $isCalledFromOverridingMethod = isset($trace[1]['function'], $trace[1]['object']) && 'serialize' === $trace[1]['function'] && $this === $trace[1]['object'];
-        }
-
-        return $isCalledFromOverridingMethod ? $serialized : serialize($serialized);
-    }
-
-    /**
      * Restores the object state from an array given by __serialize().
      *
      * There is no need to unserialize any entry in $data, they are already ready-to-use.
@@ -99,48 +80,6 @@ class AuthenticationException extends RuntimeException
     }
 
     /**
-     * {@inheritdoc}
-     *
-     * @final since Symfony 4.3, use __unserialize() instead
-     *
-     * @internal since Symfony 4.3, use __unserialize() instead
-     */
-    public function unserialize($serialized)
-    {
-        $this->__unserialize(\is_array($serialized) ? $serialized : unserialize($serialized));
-    }
-
-    /**
-     * @internal
-     */
-    public function __sleep()
-    {
-        if (__CLASS__ !== $c = (new \ReflectionMethod($this, 'serialize'))->getDeclaringClass()->name) {
-            @trigger_error(sprintf('Implementing the "%s::serialize()" method is deprecated since Symfony 4.3, implement the __serialize() and __unserialize() methods instead.', $c), E_USER_DEPRECATED);
-            $this->serialized = $this->serialize();
-        } else {
-            $this->serialized = $this->__serialize();
-        }
-
-        return ['serialized'];
-    }
-
-    /**
-     * @internal
-     */
-    public function __wakeup()
-    {
-        if (__CLASS__ !== $c = (new \ReflectionMethod($this, 'unserialize'))->getDeclaringClass()->name) {
-            @trigger_error(sprintf('Implementing the "%s::unserialize()" method is deprecated since Symfony 4.3, implement the __serialize() and __unserialize() methods instead.', $c), E_USER_DEPRECATED);
-            $this->unserialize($this->serialized);
-        } else {
-            $this->__unserialize($this->serialized);
-        }
-
-        unset($this->serialized);
-    }
-
-    /**
      * Message key to be used by the translation component.
      *
      * @return string
@@ -158,5 +97,24 @@ class AuthenticationException extends RuntimeException
     public function getMessageData()
     {
         return [];
+    }
+
+    /**
+     * @internal
+     */
+    public function __sleep()
+    {
+        $this->serialized = $this->__serialize();
+
+        return ['serialized'];
+    }
+
+    /**
+     * @internal
+     */
+    public function __wakeup()
+    {
+        $this->__unserialize($this->serialized);
+        unset($this->serialized);
     }
 }
